Compatibility with web book and viz
The changes introduced since the fork have removed some globals from the
export, so that the visualization doesn't work anymore.

Here those are restored, and the new functions that were added are
scoped elsewhere, to keep things clean.

I also pinned the dependencies to versions that do not require patching.

However, one problem remains: the satisfia mdp constructor returns an
object that doesn't have all the fields that a world and an mdp usually
have in the webppl world.
I tried returning somerthing like `_.assign({ new fields },
original_object)`, but when I do that computations fail (In computing
`V` the library complains that an element of a vector is negative).

We should find out why that is, and return a proper object from
`VerySimpleGW`. Right now I went around it by returning also the
original `world`, so that the `draw` functions knows what to paint.